# api/openapi/techstack-v1.yaml
openapi: 3.1.0
info:
  title: kombify-TechStack API
  description: |
    REST API for kombify-TechStack -- the hybrid cloud control plane for homelabs.

    kombify-TechStack manages infrastructure through a spec-driven architecture:
    validate YAML specs via CUE, generate OpenTofu configurations, and manage
    nodes through a gRPC agent protocol.

    ## Authentication

    Authenticated endpoints accept an OIDC/JWKS bearer token or an API key
    where documented. Self-hosted TechStack defaults to Pocket ID as identity
    provider. SaaS routes enter through the edge with Auth0 as the platform
    identity provider. PocketBase session exchange remains legacy migration
    compatibility only.

    ## Response Format

    **Success:** `{ "data": <object>, "meta": { "total": int, "page": int, "per_page": int } }`

    **Error:** `{ "error": { "code": "<ERROR_CODE>", "message": "...", "details": <any> } }`

    ## Pagination

    List endpoints accept `page` (default: 1) and `per_page` (default: 20, max: 100) query parameters.
  version: 1.0.0
  contact:
    name: kombify
    url: https://kombify.io
  license:
    name: Apache-2.0 AND GPL-3.0
    url: https://github.com/kombifyio/TechStack/blob/main/LICENSE

# ============================================================
# SERVERS
# ============================================================
# All paths in this document are authored as /api/v1/*. The Gateway projects
...
servers:
  - url: https://api.kombify.io/v1/techstack
...

...
security:
  - bearerAuth: []
paths:
...
  /api/v1/servers/{serverId}/ports:
    get:
      operationId: getServerRuntimePorts
      tags: [Servers, Inventory]
      summary: Get the authenticated owner's port inventory for one canonical server
      description: >-
        Server-detail UI projection over the same central port authority as the
        Inventory/MCP surface, authorized through the canonical owner-scoped
        server read boundary.
      parameters:
        - $ref: "#/components/parameters/serverId"
      responses:
        "200":
          description: Owner-scoped port allocation inventory
          headers:
            Cache-Control:
              schema: { type: string, example: "private, no-store" }
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: "#/components/schemas/ServerPortInventory" }
                  meta: { $ref: "#/components/schemas/ResponseMeta" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
  /api/v1/servers:
    get:
      operationId: listServerRuntimes
      tags: [Servers]
      summary: List canonical Node runtime state
      description: |
        Returns owner-scoped Nodes with canonical `node_id` and
        `kit_deployment_id` identities.
      parameters:
        - name: kit_deployment_id
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Owner-scoped server runtime list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ServerRuntime"
                  meta:
                    $ref: "#/components/schemas/ResponseMeta"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/servers/{serverId}:
    get:
      operationId: getServerRuntime
      tags: [Servers]
      summary: Get canonical server lifecycle, connection, health, and channels
      parameters:
        - $ref: "#/components/parameters/serverId"
      responses:
        "200":
          description: Server runtime
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ServerRuntime"
                  meta:
                    $ref: "#/components/schemas/ResponseMeta"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/servers/{serverId}/transitions:
    get:
      operationId: listServerRuntimeTransitions
      tags: [Servers]
      summary: List the persisted server state timeline
      parameters:
        - $ref: "#/components/parameters/serverId"
      responses:
        "200":
          description: Newest-first transition timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /api/v1/servers/{serverId}/detach:
    post:
      operationId: detachSelfOwnedServer
      tags: [Servers]
      summary: Revoke and detach one exact customer-operated BYO server
      description: Retains the terminal canonical server receipt and performs no provider API call.
      parameters:
        - $ref: "#/components/parameters/serverId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelfOwnedServerDetachRequest"
      responses:
        "200":
          description: Terminal BYO detach receipt
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/SelfOwnedServerDetachReceipt"
                  meta:
                    $ref: "#/components/schemas/ResponseMeta"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # ============================================================
  # SERVICE RUNTIME REGISTRY
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC/JWKS bearer token. Self-hosted defaults to Pocket ID; SaaS uses Auth0 through the edge. PocketBase JWT is legacy migration compatibility only.
    runtimeAgentAuth:
      type: http
      scheme: bearer
      bearerFormat: TechStack-Runtime-Agent
      description: Runtime-agent credential bound to the exact tenant, placement, server, and worker identity. Signed, opaque, and explicit legacy compatibility tokens are validated fail-closed by the worker runtime.
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service accounts and automation clients (SaaS mode via the public edge)
    edgeTrustHeader:
      type: apiKey
      in: header
      name: X-Edge-Internal-Trust
      description: Legacy trust header injected by the public edge during the migration cutover (SaaS mode only)
  parameters:
    serverId:
      name: serverId
      in: path
      required: true
      description: Canonical server runtime ID
      schema:
        type: string
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Request conflicts with existing state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceUnavailable:
      description: Required control-plane persistence or dependency is unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ServerPortInventory:
      type: object
      required: [server_id, server_generation, inventory_revision, listeners_complete, exposures_complete, allocations]
      properties:
        server_id: { type: string }
        server_generation: { type: integer, format: int64, minimum: 1 }
        observed_at:
          oneOf:
            - { type: string, format: date-time }
            - { type: "null" }
        expires_at:
          oneOf:
            - { type: string, format: date-time }
            - { type: "null" }
        inventory_revision: { type: integer, format: int64, minimum: 0 }
        listeners_complete: { type: boolean }
        exposures_complete: { type: boolean }
        allocations: { type: array, items: { $ref: "#/components/schemas/ServerPortAllocation" } }
...
    ServerRuntime:
      type: object
      required: [id, node_id, name, lifecycle, connection, health, channels, inventory_revision, provider, mutations_allowed, allowed_actions, stack_actions, created_at, updated_at]
      properties:
        id:
          type: string
          description: Compatibility alias for node_id.
        node_id:
          type: string
        kit_deployment_id:
          type: string
        worker_id:
          type: string
        name:
          type: string
        lifecycle:
          type: object
          required: [state, desired_state]
          properties:
            state:
              type: string
              enum: [planned, provisioning, enrolling, active, failed, decommissioning, decommissioned]
            desired_state:
              type: string
              enum: [running, stopped, absent]
            ended_at:
              type: string
              format: date-time
        connection:
          type: object
          required: [state, changed_at]
          properties:
            state:
              type: string
              enum: [pending, connecting, connected, degraded, stale, offline, revoked]
            reason_code:
              type: string
            changed_at:
              type: string
              format: date-time
            last_heartbeat_at:
              type: string
              format: date-time
            staleness_seconds:
              type: integer
              format: int64
              minimum: 0
        health:
          type: object
          required: [state]
          properties:
            state:
              type: string
              enum: [unknown, healthy, degraded, unhealthy]
            observed_at:
              type: string
              format: date-time
        channels:
          type: array
          items:
            $ref: "#/components/schemas/ServerChannel"
        inventory_revision:
          type: integer
          format: int64
          minimum: 0
        provider:
          type: object
          properties:
            lease_id:
              type: string
            ref:
              type: string
        environment_class:
          type: string
          enum: [local, cloud, unknown]
        offering:
          type: string
          enum: [self_owned_device, external_vps, managed_vps]
        provider_id:
          type: string
        provider_target_ref:
          type: string
        availability_owner:
          type: string
          enum: [customer, provider]
        operations_owner:
          type: string
          enum: [customer, kombify]
        target_evidence:
          type: object
          required: [freshness]
          properties:
            ref:
              type: string
            observed_at:
              type: string
              format: date-time
            freshness:
              type: object
              required: [state]
              properties:
                state:
                  type: string
                  enum: [recorded, unknown]
                age_seconds:
                  type: integer
                  format: int64
                  minimum: 0
        mutations_allowed:
          type: boolean
          description: Coarse legacy gate. `allowed_actions` is the precise contract.
        allowed_actions:
          type: array
          uniqueItems: true
          description: >-
            Node-scoped operations this server's own state admits. Each id maps
            to exactly one endpoint that exists; an unknown lifecycle advertises
            nothing rather than guessing that a mutation is safe. The
            lease-scoped ids target `provider.lease_id`. There is deliberately
            no node `logs` id - no such endpoint exists.
          items:
            type: string
            enum: [connect, terminal, reconnect, rotate_credentials, ssh_enable, ssh_disable, detach, decommission, resolve_custody]
        stack_actions:
          type: array
          uniqueItems: true
          description: >-
            StackKit-deployment operations, all routed to
            POST /api/v1/stacks/{id}/stackkit/operations. Kept separate from
            `allowed_actions`: a node operation and a kit operation have
            different blast radius and different authority.
          items:
            type: string
            enum: [plan, apply, verify, upgrade, drift_detect, drift_reconcile]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SelfOwnedServerDetachRequest:
      type: object
      additionalProperties: false
      required: [confirm_server_id]
      properties:
        confirm_server_id:
          type: string
          minLength: 1
          maxLength: 256
          description: Must exactly equal the serverId path parameter.
    SelfOwnedServerDetachReceipt:
      type: object
      required: [server_id, agent_id, revision, generation, detached_at, replay]
      properties:
        server_id:
          type: string
        agent_id:
          type: string
        revision:
          type: integer
          format: int64
          minimum: 1
        generation:
          type: integer
          format: int64
          minimum: 1
        detached_at:
          type: string
          format: date-time
        replay:
          type: boolean
...
    SuccessEnvelope:
      type: object
      properties:
        data:
          description: Response payload
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - INTERNAL_ERROR
                - NOT_FOUND
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - CONFLICT
                - VALIDATION_ERROR
                - RATE_LIMITED
                - SERVICE_UNAVAILABLE
                - METHOD_NOT_ALLOWED
            message:
              type: string
            details:
              description: Additional error context (validation errors, etc.)
    ResponseMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        per_page:
          type: integer
        managed_runtime_inventory_complete:
          type: boolean
          description: True only when the managed RuntimeLease authority was configured and read successfully for this inventory response.
        worker_inventory_sha256:
          type: string
          pattern: '^[0-9a-f]{64}$'
          description: Stable digest of the complete worker identity, attachment, freshness, and managed RuntimeLease authority snapshot used for pagination.
        managed_runtime_active_lease_ids:
          type: array
          items:
            type: string
          description: Canonically sorted IDs of visible managed RuntimeLeases classified as active for inventory and cleanup. This visibility projection is not capacity reservation or release authority.
        managed_runtime_inactive_lease_ids:
          type: array
          items:
            type: string
          description: Canonically sorted IDs of visible managed RuntimeLeases classified as cancelled or archived for inventory and cleanup. Inactive state does not prove provider absence or release managed-server capacity.
        managed_runtime_lease_generation_digests:
          type: array
          items:
            type: object
            required: [lease_id, resource_generation_digest]
            properties:
              lease_id:
                type: string
              resource_generation_digest:
                type: string
                pattern: '^[0-9a-f]{64}$'
          description: Canonically sorted opaque bindings from each visible managed RuntimeLease ID to its exact provider-resource generation. Missing coverage makes destructive inventory proof invalid while leaving ordinary dashboard inventory readable.
        managed_runtime_duplicate_lease_ids:
          type: array
          items:
            type: string
          description: Managed RuntimeLease IDs represented by multiple persisted worker rows before UI projection deduplication; must be empty for destructive inventory proof.
        managed_runtime_attachment_conflict_lease_ids:
          type: array
          items:
            type: string
          description: Canonically sorted caller-visible managed RuntimeLease IDs with a conflicting tenant-wide worker attachment; must be empty for destructive inventory proof and never exposes foreign private lease or worker details.
...